move files to screen folder
